Allows ms_run[i]_location to be "null" as placeholder if the actual location is unknown (mzTab specification v1.0.0. section 6.2.35)
Update issue 13
Status: Fixed
The errors reported previously in the parser have been transformed as warnings. Previously the parser stopped if an error was found in the metadata. As the warnings are stored internally in the same error list than the errors (something to improve in the future) the behaviour of the parser has been changed and it will not stop if an error or a warning it is found in the metadata section, it will try to continue parsing the rest of the document.