New versions of the Progenesis examples
Changing the evidence_grouping_ID column to reflect the unique
evidence group used to support an identification. Within Progenesis,
this is always an aggregate of all the MS or MS/MS data from the
deconvoluted features, so the evidence_grouping_ID values actually just
map back to the SML_ID. Which is convenient, though it won't hold true
outside Progenesis.

Also correct the mixup on whether assay[i] or ms_run[i] was titled,
correcting issue #44.